Separate launcher's matrix locations in to a separate aggregate. Add functions to shift launchers / make slots available in preparation for a more versatile AddLauncher implementation.
