+++ desktop/src/aggregates/mod.rs
@@ -1,5 +1,7 @@
 pub mod layout_specs;
+pub mod matrix_positions;
 pub mod ordered_hierarchy;
 
 pub use layout_specs::*;
+pub use matrix_positions::*;
 pub use ordered_hierarchy::*;
