+++ desktop/src/desktop_system/layout_effects.rs
@@ -46,7 +46,8 @@ impl DesktopSystem {
 
     pub(super) fn apply_focused_view_window_state(&self) -> Result<()> {
         let state = self.focused_view_window_state()?.unwrap_or_default();
-        self.window.set_title(&self.focused_window_title(state.title)?);
+        self.window
+            .set_title(&self.focused_window_title(state.title)?);
         self.window.set_cursor(state.cursor);
         // Pointer-feedback state drives cursor visibility (hidden during keyboard navigation).
         self.window.set_cursor_visible(self.is_cursor_visible());
@@ -75,7 +76,7 @@ impl DesktopSystem {
             });
 
         let mut title = if terminal_title.is_empty() {
-                self.env.primary_application.clone()
+            self.env.primary_application.clone()
         } else {
             terminal_title
         };
