+++ desktop/src/projects/launcher_presenter.rs
@@ -19,7 +19,7 @@ use super::visor_layout;
 use crate::Map;
 use crate::desktop_system::{Commands, DesktopCommand, place_container_children};
 use crate::instance_presenter::InstancePresenter;
-use crate::projects::{LaunchProfileId, MatrixPlacement};
+use crate::projects::LaunchProfileId;
 
 use super::configuration::{LaunchProfile, LauncherMode};
 
@@ -49,7 +49,6 @@ pub struct LauncherPresenter {
     #[allow(unused)]
     id: LaunchProfileId,
     profile: LaunchProfile,
-    pub placement: MatrixPlacement,
     mode: LauncherMode,
     layout_transform: Transform,
     scene_transform: Handle<Transform>,
@@ -78,7 +77,6 @@ impl LauncherPresenter {
     pub fn new(
         parent_location: Handle<Location>,
         id: LaunchProfileId,
-        placement: MatrixPlacement,
         profile: LaunchProfile,
         size: Size,
         scene: &Scene,
@@ -116,7 +114,6 @@ impl LauncherPresenter {
         Self {
             id,
             profile,
-            placement,
             mode,
             layout_transform: Transform::IDENTITY,
             scene_transform: our_transform,
